django/boardcast/requirements.txt
Django==5.0.10
djangorestframework==3.15.2
django-environ==0.11.2
channels==4.1.0
channels-redis==4.2.0
django-cors-headers==4.4.0

# Worker (optional but recommended)
celery==5.4.0
redis==5.0.8

# For uploads / validation
python-magic==0.4.27; platform_system != "Windows"
